Improved navigation at mainPage
window.location is a bad standard for redirecting as browsers struggle to recall the past location. I wanted to use a hyperlink for the ComposeTask button, however, I wanted to avoid the duplication of data. Since then, I was able to learn more about React, and figured to use a spread operator as attributes instead, which helps me eliminate the cluster of duplicate jsx props (attributes).