Updated documentation plus iteration mistakes and worked on composeTask page
More iteration in terms of design. I forgot to include two issues related to composing task. More can be found in updated documentation.

I have also updated the create task functionality at Compose-Task page. User is able to save valid task, however, I had to stop due to the designs flaws, hence my redirected focus towards required iteration. All I have to do is, redirect user with the newly created task to the Detailed-Task page.

Then redirect focus on other pages.

Also here to document potential improvements for future evaluation. LocalSotrage is not ideal because I often have to call all saved tasks and data is at high risk since client is able to clear their cache. It would be ideal to store data in a database but we may require addition of a backend. The purpose of this project is to learn React, hence I will just stick to the localStorage and explain myself later in areas for improvement.