Fixing bugs & finished taskDetail page, improved loginComponent
loginComponent is now editable thanks to the message prop. It is so that we can clearly indicate the pages' actions.

User is now able to delete their tasks, and is redirected back to the main page once delete-action is successful.

Composing task had an issue with assigning id's since it relied on the arrays length. However, length changes. If we delete a middle task. The next tasks length does not correspond to their id. I have fixed this issue by identifying and then assigning biggest id for the newly created task.